マウス操作によってクリックしたらオブジェクトを置けるようになった。
入力関連の関数名を変えたので注意！(keyboard_update→input_update,keyboard_erase→input_erase)
またinput.hにマウスの入力情報を取得できるようにしました。

+++ SSR_Collaboration_BU/SSR_Collaboration_BU/Terrain.cpp
@@ -2,7 +2,7 @@
 #include"Terrain.h"
 
 
-Terrain::Terrain(const std::shared_ptr<MyShape> hitJudgeShape, double x, double y, int gHandle, int kind, unsigned int color, bool breakable)
+Terrain::Terrain(const std::shared_ptr<MyShape> hitJudgeShape, float x, float y, int gHandle, int kind, unsigned int color, bool breakable)
 	:BattleObject(hitJudgeShape,x,y,gHandle,kind),m_color(color),m_breakable(breakable){}
 
 Terrain::~Terrain() {}
